Built Input Class
Added a few lines to the driver to process the lines and feed them into
the Input constructor.

Built the Input class to fully parse any input and output an error for
any bad inputs.

Added a test file with a range of test cases.

Currently, the client only tells you a valid Input object was created.
If one wasn’t created, an error message is given in the Input class.