Updated classes
Added the overwritten calculatePrice and printItemAttributes methods in
each of the three classes and tested items to print them out.  The item
testing is shown at the top of the driver class.

All that really remains is the logic to control the arrayList and
general testing. arrayList functionality should probably be implemented
in the order of:
insert
print
search
modify insert to check for existing object first
update
delete