Methods updated
Search finds and reports the quantity, name, and price of all objects
with the requested name or an error message.

print displays information about every item in the cart and posts the
total cart value.  To facilitate this, printItemAttributes now returns
a double.  This keeps calculatePrice as a private method.

Insert currently handles any object with a new name.  Any additional
input with the same name is currently ignored.  This needs to be
changed to insert the item if it is different than the current item or
consolidate otherwise.  To change this functionality to insert all
items, just need to comment out the line making alreadyExists = true.
This can be done if we can’t figure out a better way to solve this.

update and delete haven’t been implemented at all to this point.