Should be the final implementation
Cleaned up excess code.
Added a reader.close() to get rid of the warning.
Added comments everywhere.

Added several additional test cases.  Here’s what we test for:
All operations to an empty cart.
Inserting an invalid category.
Printing with additional input
Searching for an item that exists and an item that doesn’t.
Updating an item that exists and an item that doesn’t.
Updating an item to a quantity of 0.
delete an item that exists and doesn’t exist
delete with additional input
an invalid operation
inserting all three category types correctly
weights that are invalid due to negative or nonnumeric values
quantities that aren’t integer values
0 quantity insert
negative quantity insert
negative price
price with multiple periods
price that starts with a period
price that doesn’t include a period
price that has 3 decimal places
electronics with an invalid fragile code
electronics with an invalid state
grogery with invalid perishable code
input with extra spaces at the end

Can you think of anything else?